refactor: ArtworkList에서 paymentType만 넘겨받도록 로직 수정
